chore: moved tests to correct file
